feat(layout): Unify auth code with main home page code
- Navbar and ProfileDropdown now are components instead of single page.
- Header handle session and show the components if user is logged in.
- Adjusted English default string for logout to match original SW360.
