+++ public/locales/en/common.json
@@ -3,7 +3,7 @@
   "Create Account": "Create Account",
   "Email Address": "Email Address",
   "In order to go ahead, please sign in or create a new account!": "In order to go ahead, please sign in or create a new account!",
-  "Logout": "Logout",
+  "Logout": "Sign Out",
   "Password": "Password",
   "Sign In": "Sign In",
   "SW360 is an open source software project that provides both a web application and a repository to collect, organize and make available information about software components. It establishes a central hub for software components in an organization.": "SW360 is an open source software project that provides both a web application and a repository to collect, organize and make available information about software components. It establishes a central hub for software components in an organization.",
